refactor: merge the socket service with the hook
