+++ src/Pages/SegmentSelectionPage/SegmentSelectionPage.tsx
@@ -6,7 +6,7 @@ import { ThumbUp, ThumbDown } from '@mui/icons-material'
 import { Button } from '@mui/material'
 
 import { CenterdContainer } from '../../components'
-import { useSocketService } from '../../services'
+import { useSocketService } from '../../hooks'
 
 const StyledButtonsContainer = styled.div`
   display: flex;
