+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useDoubleKeyPress'
 export * from './useProgram'
 export * from './useSupabase'
+export * from './useSocketService'
