--- src/services/SocketService/Socket.service.ts
@@ -1,110 +1,0 @@
-import { io, Socket } from 'socket.io-client'
-
-import { EventNames } from './eventNames'
-import { ProgramType } from '../../types'
-
-export class SocketService {
-  private _socket: Socket
-
-  constructor(port = 8000) {
-    this._socket = io('http://localhost:' + port)
-    this._socket.emit(EventNames.CLIENT_CONNECTED, 'A new client connected')
-  }
-
-  public onStart = (exicute: () => void) => {
-    this._socket.on(EventNames.START_PROGRAM, (data: any) => {
-      // eslint-disable-next-line no-console
-      console.log(
-        `Received command (start-program): ${data}`,
-        new Date().getMilliseconds()
-      )
-      exicute()
-    })
-  }
-  public onPause = (exicute: () => void) => {
-    this._socket.on(EventNames.PAUSE_PROGRAM, (data: any) => {
-      // eslint-disable-next-line no-console
-      console.log(
-        `Received command (pause-program): ${data}`,
-        new Date().getMilliseconds()
-      )
-      exicute()
-    })
-  }
-  public onReset = (exicute: () => void) => {
-    this._socket.on(EventNames.RESET_PROGRAM, (data: any) => {
-      // eslint-disable-next-line no-console
-      console.log(
-        `Received command (reset-program): ${data}`,
-        new Date().getMilliseconds()
-      )
-      exicute()
-    })
-  }
-  public onRequestFullScreen = (exicute: () => void) => {
-    this._socket.on(EventNames.REQUEST_FULLSCREEN, (data: any) => {
-      // eslint-disable-next-line no-console
-      console.log(
-        `Received command (request-full-screen): ${data}`,
-        new Date().getMilliseconds()
-      )
-      exicute()
-    })
-  }
-  public onSetProgram = (exicute: (program: ProgramType) => void) => {
-    this._socket.on(EventNames.SET_PROGRAM, (program: ProgramType) => {
-      // eslint-disable-next-line no-console
-      console.log(
-        `Received command (set-program): `,
-        new Date().getMilliseconds()
-      )
-      exicute(program)
-    })
-  }
-
-  public onShowControls = (exicute: () => void) => {
-    this._socket.on(EventNames.SHOW_CONTROLS, () => {
-      // eslint-disable-next-line no-console
-      console.log(
-        `Received command (show-controls): `,
-        new Date().getMilliseconds()
-      )
-      exicute()
-    })
-  }
-  public onHideControls = (exicute: () => void) => {
-    this._socket.on(EventNames.HIDE_CONTROLS, () => {
-      // eslint-disable-next-line no-console
-      console.log(
-        `Received command (hide-controls): `,
-        new Date().getMilliseconds()
-      )
-      exicute()
-    })
-  }
-  public onUserSelectedNextSegment = (
-    exicute: (selectedNextSegmentIndex: number) => void
-  ) => {
-    this._socket.on(
-      EventNames.USER_SELECTED_SEGMENT,
-      (selectedNextSegmentIndex: number) => {
-        // eslint-disable-next-line no-console
-        console.log(
-          `Received command (user-selected-segment): `,
-          new Date().getMilliseconds()
-        )
-        exicute(selectedNextSegmentIndex)
-      }
-    )
-  }
-  public onToggleShowControls = (exicute: () => void) => {
-    this._socket.on(EventNames.TOGGLE_SHOW_CONTROLS, () => {
-      // eslint-disable-next-line no-console
-      console.log(
-        `Received command (toggle-show-controls): `,
-        new Date().getMilliseconds()
-      )
-      exicute()
-    })
-  }
-}
