--- src/services/SocketService/index.ts
@@ -1,3 +1,0 @@
-export * from './Socket.service'
-export * from './eventNames'
-export * from './useSocketService'
