+++ src/services/index.ts
@@ -1,3 +1,2 @@
 export * from './ScreenService'
 export * from './VideoService'
-export * from './SocketService'
